+++ oap-cache/oap/src/main/java/org/apache/spark/sql/execution/datasources/oap/filecache/PlasmaTimeOutWrapper.java
@@ -1,3 +1,20 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.apache.spark.sql.execution.datasources.oap.filecache;
 
 import java.util.concurrent.*;
@@ -15,14 +32,14 @@ public class PlasmaTimeOutWrapper implements Callable<Object> {
   private ExecutorService executorService;
   private Object[] paramValues;
   private long timeOutInSeconds;
-  private PlasmaLambdaWrapper wrapper;
+  private PlasmaMethodWrapper wrapper;
   private PlasmaClient client;
 
   private PlasmaTimeOutWrapper() {
   }
 
   public static Object run(
-          PlasmaLambdaWrapper wrapper,
+          PlasmaMethodWrapper wrapper,
           PlasmaClient client,
           ExecutorService executorService,
           Object[] paramValues, long timeOutInSeconds)
@@ -34,7 +51,7 @@ public class PlasmaTimeOutWrapper implements Callable<Object> {
   }
 
   private Object submitFutureTask(
-          PlasmaLambdaWrapper wrapper,
+          PlasmaMethodWrapper wrapper,
           PlasmaClient client,
           ExecutorService executorService,
           Object[] paramValues, long timeOutInSeconds)
